Remove unused helper methods; extracts useful ones to new centralised test_helper file; test_helper included in spec_helper.
