+++ spec/bot/intents/classifier_spec.rb
@@ -1,7 +1,6 @@
 require 'rails_helper'
 
 RSpec.describe Intents::Classifier do
-  FakeMessage = Struct.new(:sender, :recipient, :timestamp, :text, :messaging)
   describe '.classify' do
     it 'correctly classifies a message asking to create an account' do
       text = 'I\'d like to create an account please'
@@ -35,25 +34,5 @@ RSpec.describe Intents::Classifier do
 
       expect(intent).to eq(:just_anything_for_now)
     end
-  end
-
-  private
-
-  def fake_incoming_message(message_text, quick_reply_payload = nil)
-    sender = {"id"=>"1234"}
-    recipient = {"id"=>"5678"}
-    timestamp = 1528049653543
-    messaging = {
-      "sender"=>{"id"=>"1732016540208841"},
-        "recipient"=>{"id"=>"364376550736984"},
-        "timestamp"=>1535288528163,
-        "message"=>{
-          "mid"=> "0j3SQeNnpIDLxBwJl7hRofACd_36bJGN3qXKXv32Bok8GqJfA284e1hsnOagFiVZsbqLLainVGIVURWOlNJ4Tw",
-          "seq"=>2171281,
-          "text"=>"Account"
-        }
-      }
-    messaging['message']['quick_reply'] = {'payload' => quick_reply_payload} if quick_reply_payload
-    FakeMessage.new(sender, recipient, timestamp, message_text, messaging)
   end
 end
