+++ spec/spec_helper.rb
@@ -15,6 +15,7 @@
 # See http://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
 require 'webmock/rspec'
 require 'factory_bot'
+require_relative './helpers/test_helper'
 
 RSpec.configure do |config|
   # rspec-expectations config goes here. You can use an alternate
